Improved the way queues are consumed
