+++ src/main/java/com/kylerdeggs/javaconnected/service/PostService.java
@@ -8,7 +8,6 @@ import com.kylerdeggs.javaconnected.web.PostDto;
 import org.apache.tika.mime.MimeTypeException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -18,7 +17,6 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.validation.constraints.NotNull;
 import java.io.IOException;
 import java.io.Serializable;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
@@ -27,7 +25,7 @@ import java.util.Optional;
  * Provides methods for retrieving, creating, updating, and deleting a post.
  *
  * @author Kyler Deggs
- * @version 1.0.0
+ * @version 1.1.0
  */
 @Service
 public class PostService {
@@ -103,6 +101,37 @@ public class PostService {
     }
 
     /**
+     * Saves a post.
+     *
+     * @param post Post to save
+     */
+    public void savePost(Post post) {
+        postRepository.save(post);
+    }
+
+    /**
+     * Sends a request to the AWS service to process an upload.
+     *
+     * @param authorId ID of the author
+     * @param media    Media that needs to be uploaded
+     * @return Key of the uploaded media
+     * @throws IOException       Media file processing error
+     * @throws MimeTypeException Trying to upload a restricted file type
+     */
+    public String saveMedia(String authorId, MultipartFile media) throws IOException, MimeTypeException {
+        return aws.processUpload(authorId, media);
+    }
+
+    /**
+     * Sends a request to the AWS service to delete media.
+     *
+     * @param key Location of the media
+     */
+    public void deleteMedia(String key) {
+        aws.deleteMedia(key);
+    }
+
+    /**
      * Processes a post creation request by sending the post to the correct RabbitMQ queue.
      *
      * @param postDto Post to be created
@@ -113,7 +142,7 @@ public class PostService {
     public void processPost(PostDto postDto, MultipartFile media) throws IOException, MimeTypeException {
         if (userService.userExists(postDto.getAuthorId())) {
             if (media != null && !media.isEmpty())
-                postDto.setMediaLocation(aws.processUpload(postDto.getAuthorId(), media));
+                postDto.setMediaLocation(saveMedia(postDto.getAuthorId(), media));
 
             LOGGER.info("A new post is being sent to the exchange " + exchangeName
                     + " to be routed to the queue " + postQueueName);
@@ -164,68 +193,12 @@ public class PostService {
     }
 
     /**
-     * Creates new posts by consuming the post creation queue.
-     *
-     * @param postDto Post to be created
-     */
-    @RabbitListener(queues = "${amqp.queue.post-name}")
-    private void postCreator(PostDto postDto) {
-        User author = userService.verifyUser(postDto.getAuthorId());
-        String mediaLocation = postDto.getMediaLocation();
-        Post post = new Post(author, postDto.getTitle(), mediaLocation != null,
-                mediaLocation, postDto.getCaption(), true, LocalDateTime.now());
-
-        postRepository.save(post);
-        LOGGER.info("A new post with ID " + post.getId() + " has been created");
-    }
-
-    /**
-     * Modifies the like count of post by consuming the like queue.
-     *
-     * @param likeDto Like that contains information on whether to add or remove a like
-     */
-    @RabbitListener(queues = "${amqp.queue.like-name}")
-    private void likeModifier(LikeDto likeDto) {
-        Post targetPost = verifyPost(likeDto.getPostId());
-
-        if (likeDto.isAddLike())
-            targetPost.setLikeCount(targetPost.getLikeCount() + 1);
-        else
-            targetPost.setLikeCount(targetPost.getLikeCount() != 0 ? targetPost.getLikeCount() - 1 : 0);
-
-        postRepository.save(targetPost);
-        LOGGER.info("A like has been " + (likeDto.isAddLike() ? "added" : "removed")
-                + " to post " + targetPost.getId());
-    }
-
-    /**
-     * Deletes posts by consuming the post deletion queue.
-     *
-     * @param postId Post to be deleted
-     */
-    @RabbitListener(queues = "${amqp.queue.post-delete-name}")
-    private void postDeleter(long postId) {
-        Post targetPost = verifyPost(postId);
-
-        if (targetPost.getHasMedia()) {
-            aws.deleteMedia(targetPost.getMediaLocation());
-            targetPost.setHasMedia(false);
-            targetPost.setMediaLocation(null);
-        }
-        targetPost.setPublished(false);
-        targetPost.setDeletedAt(LocalDateTime.now());
-
-        postRepository.save(targetPost);
-        LOGGER.info("Post " + postId + " has been unpublished");
-    }
-
-    /**
      * Like representation object.
      *
      * @author Kyler Deggs
      * @version 1.0.0
      */
-    private static class LikeDto implements Serializable {
+    public static class LikeDto implements Serializable {
         @NotNull
         private final long postId;
